Fix: an incorrect constructor was used in the Saver, leading to incorrect color selection after e.g. screen rotation

+++ core/src/commonMain/kotlin/dev/zt64/compose/pipette/HsvColor.kt
@@ -108,7 +108,7 @@ public value class HsvColor private constructor(public val packedValue: ULong) {
     public companion object {
         public val Saver: Saver<HsvColor, Long> = Saver(
             save = { it.packedValue.toLong() },
-            restore = { HsvColor(it) }
+            restore = { HsvColor(it.toULong()) }
         )
 
         private fun hsvToRgbComponent(n: Int, h: Float, s: Float, v: Float): Float {
